Move SQL querier and types converter to HSQLDB storage
